pyproject.toml
[project]
name = "ghostty-bot"
version = "0.1.0"
description = ""
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
  { name = "Mitchell Hashimoto", email = "m@mitchellh.com" },
  { name = "trag1c", email = "dev@jakubr.me" },
]
dependencies = [
  "discord-py~=2.6.3",
  "githubkit~=0.13.1",
  "pydantic>=2.11.7,<3",
  "sentry-sdk~=2.36",
  "zig-codeblocks==0.3.3",
  "httpx~=0.28.1",
  "monalisten==0.2.0",
  "pydantic-settings>=2.10.1,<3",
  "loguru~=0.7.0",
]

[dependency-groups]
dev = [
  "ruff==0.12.10",
  "basedpyright==1.31.4",
  "taplo==0.9.3",
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
]

[tool.pyright]
reportAny = false
reportExplicitAny = false
reportImplicitStringConcatenation = false
reportInvalidCast = false
reportUnusedCallResult = false
reportUnknownLambdaType = false

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = ["ignore::Warning"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM", "ERA", "D", "FIX", "S", "ANN401", "ISC001", "TD003", "PLR2004"]
mccabe.max-complexity = 20
pylint.max-branches = 20
pylint.max-returns = 15
isort.known-local-folder = ["app"]

[tool.ruff.format]
preview = true

[tool.ruff.lint.flake8-import-conventions.aliases]
discord = "dc"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["FBT", "SLF001"]
"./app/components/github_integration/__init__.py" = ["PLC0415"]
